random fixes

+++ bioinformatic algorithms/hw2/problem3_bloom_filter/hash_functions.cpp
@@ -5,21 +5,21 @@
 void HashFunction::generateCoefficients() {
   coefficients.clear();
   srand(time(NULL));
-  for (int i = 0; i < _keys_count; i++) {
+  for (size_t i = 0; i < _keys_count; i++) {
     coefficients.push_back(rand() % _modulo);
   }
 }
 
 void HashFunction::regenerateCoefficients() {
   srand(time(NULL));
-  for (int i = 0; i < _keys_count; i++) {
+  for (size_t i = 0; i < _keys_count; i++) {
     coefficients[i] = rand() % _modulo;
   }
 }
 
-int HashFunction::getValue(std::string  const & kmer) {
-  int result = 0;
-  for (int i = 0; i < static_cast<int>(kmer.size()); i++) {
+size_t HashFunction::getValue(std::string  const & kmer) {
+  size_t result = 0;
+  for (size_t i = 0; i < kmer.size(); i++) {
     result += coefficients[i]*kmer[i];
     result %= _modulo;
   }
@@ -31,7 +31,7 @@ bool HashFunction::operator ==(HashFunction const& hash) {
     return false;
 
   bool result = true;
-  for (int i = 0; i < _keys_count; i++) {
+  for (size_t i = 0; i < _keys_count; i++) {
     result &= (this->coefficients[i] == hash.coefficients[i]);
   }
   return result;
